refactor(test): split macro-based containment tests into separate module
Extract macro-generated containment tests from containment_with_index_tests.rs
into dedicated jsonb_containment_uses_index_tests.rs module.

Original hand-written tests remain in containment_with_index_tests.rs,
while macro infrastructure and generated tests move to new module.
